Update Git `shove` alias to lease refs
From the Git docs:

> Usually, "git push" refuses to update a remote ref that is not an
ancestor of the local ref used to overwrite it.

> This option overrides this restriction if the current value of the
remote ref is the expected value. "git push" fails otherwise.

> Imagine that you have to rebase what you have already published.
You will have to bypass the "must fast-forward" rule in order to
replace the history you originally published with the rebased
history. If somebody else built on top of your original history while
you are rebasing, the tip of the branch at the remote may advance
with her commit, and blindly pushing with --force will lose her work.

> This option allows you to say that you expect the history you are
updating is what you rebased and want to replace. If the remote ref
still points at the commit you specified, you can be sure that no
other people did anything to the ref. It is like taking a "lease" on
the ref without explicitly locking it, and the remote ref is updated
only if the "lease" is still valid.

> `--force-with-lease` alone, without specifying the details, will
protect all remote refs that are going to be updated by requiring
their current value to be the same as the remote-tracking branch we
have for them.